fix: req10

+++ desafio10.sql
@@ -1,9 +1,9 @@
 DELIMITER $$
 
-CREATE FUNCTION quantidade_musicas_no_historico(id TINYINT)
-RETURNS DOUBLE READS SQL DATA
+CREATE FUNCTION quantidade_musicas_no_historico(id INT)
+RETURNS INT READS SQL DATA
 BEGIN
-DECLARE resp DOUBLE;
+DECLARE resp INT;
 SELECT COUNT(musicas_id) FROM SpotifyClone.historico_reproducao
 WHERE usuario_id = id
 INTO resp;
